+++ backend/canisters/local_user_index/impl/src/updates/bot_delete_channel.rs
@@ -1,32 +1,37 @@
 use canister_api_macros::update;
 use local_user_index_canister::bot_delete_channel::*;
+use oc_error_codes::OCErrorCode;
 use types::BotActionScope;
 
 use crate::{bots::extract_access_context, mutate_state};
 
 #[update(candid = true, json = true, msgpack = true)]
 async fn bot_delete_channel(args: Args) -> Response {
-    use Response::*;
-
     let context = match mutate_state(|state| extract_access_context(&args.auth_token, state)) {
         Ok(context) => context,
-        Err(error) => return FailedAuthentication(error),
+        Err(_) => return OCErrorCode::BotNotAuthenticated.into(),
     };
 
     let community_id = match context.scope {
         BotActionScope::Chat(details) => match details.chat {
             types::Chat::Channel(community_id, channel_id) => {
                 if channel_id != args.channel_id {
-                    return InvalidRequest("Channel ID does not match access token".to_string());
+                    return OCErrorCode::InvalidRequest
+                        .with_message("Channel ID does not match access token")
+                        .into();
                 }
                 community_id
             }
-            _ => return InvalidRequest("Must be community or channel scope".to_string()),
+            _ => {
+                return OCErrorCode::InvalidBotActionScope
+                    .with_message("Must be community scope")
+                    .into();
+            }
         },
         BotActionScope::Community(details) => details.community_id,
     };
 
-    match community_canister_c2c_client::c2c_bot_delete_channel(
+    community_canister_c2c_client::c2c_bot_delete_channel(
         community_id.into(),
         &community_canister::c2c_bot_delete_channel::Args {
             channel_id: args.channel_id,
@@ -35,11 +40,5 @@ async fn bot_delete_channel(args: Args) -> Response {
         },
     )
     .await
-    {
-        Ok(response) => match response {
-            community_canister::c2c_bot_delete_channel::Response::Success => Success,
-            community_canister::c2c_bot_delete_channel::Response::Error(error) => Error(error),
-        },
-        Err(error) => Error(error.into()),
-    }
+    .into()
 }